Fix compiler warning.
git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@7815 c06c8d41-db1a-0410-9941-cceddc491573

+++ crawl-ref/source/spl-book.cc
@@ -732,7 +732,7 @@ spell_type which_spell_in_book(const item_def &book, int spl)
 spell_type which_spell_in_book(int sbook_type, int spl)
 {
     ASSERT( sbook_type >= 0 );
-    ASSERT( sbook_type < NUMBER_SPELLBOOKS );
+    ASSERT( sbook_type < static_cast<int>(NUMBER_SPELLBOOKS) );
     return spellbook_template_array[sbook_type][spl];
 }                               // end which_spell_in_book()
